chore: separate colour palette doc page
